Implement validation for a subset of the InVEST models

Approved-by: Richard Sharp <richpsharp@gmail.com>

+++ src/natcap/invest/fisheries/fisheries_model.py
@@ -138,7 +138,7 @@ def set_recru_func(vars_dict):
 
         N_next[0], spawners = rec_func(N_prev)
     '''
-    sexsp = float(vars_dict['sexsp'])
+    sexsp = int(vars_dict['sexsp'])
     LarvDisp = vars_dict['Larvaldispersal']
 
     # Initialize Weight vector according to spawn_units
